requirements.txt
numpy>=1.9.0
scipy>=0.15.0
javabridge>=1.0.0
python-bioformats>=1.0.0